Ajout d'interface pour userController et Userservice
